+++ molo/core/tests/test_views.py
@@ -1,4 +1,6 @@
+# -*- coding: utf-8 -*-
 from os import environ
+import re
 import json
 import pytest
 import responses
@@ -326,20 +328,33 @@ class TestPages(TestCase, MoloTestCaseMixin):
             title='Footer Page in french')
 
         response = self.client.get('/')
-        self.assertContains(response, 'Footer Page')
-        self.assertContains(
-            response,
-            '<a href="/footers-main-1/footer-page/">Footer Page</a>',
-            html=True)
-        self.assertNotContains(
-            response,
-            '<a href="/%s/">Footer Page in french</a>' % footer_french.slug,
-            html=True)
+
+        soup = BeautifulSoup(response.content, 'html.parser')
+        self.assertTrue(soup.findAll(
+            'a',
+            string=re.compile(self.footer.title),
+            attrs={'href': '/footers-main-1/footer-page/'}
+        ))
+
+        self.assertFalse(soup.findAll(
+            'a',
+            string=re.compile(footer_french.title),
+        ))
 
         response = self.client.get(
             '/sections-main-1/your-mind/your-mind-subsection/')
-        self.assertContains(response, 'Footer Page')
-        self.assertNotContains(response, 'Footer Page in french')
+
+        soup = BeautifulSoup(response.content, 'html.parser')
+        self.assertTrue(soup.findAll(
+            'a',
+            string=re.compile(self.footer.title),
+            attrs={'href': '/footers-main-1/footer-page/'}
+        ))
+
+        self.assertFalse(soup.findAll(
+            'a',
+            string=re.compile(footer_french.title),
+        ))
 
     def test_section_listing(self):
         self.mk_articles(
@@ -360,8 +375,8 @@ class TestPages(TestCase, MoloTestCaseMixin):
         response = self.client.get('/')
         self.assertContains(
             response,
-            '<a href="/locale/en/?next=/?" class="language-list__anchor '
-            'language-list__anchor-with-label is-active">English</a>',
+            '<a href="/locale/en/?next=/" class="language-list__anchor '
+            'language-list__anchor--with-label is-active">English</a>',
             html=True)
         self.assertContains(
             response, '<a href="/locale/fr/')
@@ -374,8 +389,8 @@ class TestPages(TestCase, MoloTestCaseMixin):
         response = client.get(self.site2.root_url)
         self.assertContains(
             response,
-            '<a href="/locale/en/?next=/?" class="language-list__anchor '
-            'language-list__anchor-with-label is-active">English</a>',
+            '<a href="/locale/en/?next=/" class="language-list__anchor '
+            'language-list__anchor--with-label is-active">English</a>',
             html=True)
         self.assertContains(
             response, '<a href="/locale/es/')
@@ -899,13 +914,13 @@ class TestPages(TestCase, MoloTestCaseMixin):
         self.assertRedirects(
             response,
             'http://main-1.localhost:8000/sections'
-            '-main-1/your-mind-in-french/?')
+            '-main-1/your-mind-in-french/')
 
         response = self.client.get('/sections-main-1/your-mind/test-page-0/')
         self.assertRedirects(
             response,
             'http://main-1.localhost:8000/sections-main-1/your-mind/'
-            'test-page-0-in-french/?')
+            'test-page-0-in-french/')
 
         # redirect from translation to main language should also work
         response = self.client.get('/locale/en/')
@@ -913,14 +928,14 @@ class TestPages(TestCase, MoloTestCaseMixin):
         response = self.client.get('/sections-main-1/your-mind-in-french/')
         self.assertRedirects(
             response,
-            'http://main-1.localhost:8000/sections-main-1/your-mind/?')
+            'http://main-1.localhost:8000/sections-main-1/your-mind/')
 
         response = self.client.get('/sections-main-1/your-mind/'
                                    'test-page-0-in-french/')
         self.assertRedirects(
             response,
             'http://main-1.localhost:8000/sections-main-1/your-mind/test-pag'
-            'e-0/?')
+            'e-0/')
 
         # unpublished translation will not result in a redirect
         self.yourmind_fr.unpublish()
@@ -1129,8 +1144,7 @@ class TestPages(TestCase, MoloTestCaseMixin):
         self.assertContains(response, '&larr;')
 
     def test_pagination_for_translated_articles_in_sections(self):
-        en_articles = self.mk_articles(self.yourmind, count=12)
-        self.mk_articles(self.yourmind, count=3)
+        en_articles = self.mk_articles(self.yourmind, count=15)
 
         for p in en_articles:
             self.mk_article_translation(
@@ -1860,3 +1874,139 @@ class TestWagtailAdmin(TestCase, MoloTestCaseMixin):
         self.assertEquals(response.status_code, 200)
         soup = BeautifulSoup(response.content, 'html.parser')
         self.assertFalse(soup.find('a', string='Explorer'))
+
+
+class TestModerationActions(TestCase, MoloTestCaseMixin):
+
+    def setUp(self):
+        self.client = Client()
+        self.mk_main()
+        self.main = Main.objects.all().first()
+        self.english = SiteLanguageRelation.objects.create(
+            language_setting=Languages.for_site(self.main.get_site()),
+            locale='en',
+            is_active=True)
+        # Creates a section under the index page
+        self.english_section = self.mk_section(
+            self.section_index, title='English section')
+        self.yourmind = self.mk_section(
+            self.section_index, title='Your mind')
+
+        self.mk_main2()
+        self.main2 = Main.objects.all().last()
+        self.language_setting2 = Languages.objects.create(
+            site_id=self.main2.get_site().pk)
+        self.english2 = SiteLanguageRelation.objects.create(
+            language_setting=self.language_setting2,
+            locale='en',
+            is_active=True)
+        self.yourmind2 = self.mk_section(
+            self.section_index2, title='Your mind')
+
+        superuser = User.objects.create_superuser(
+            username='testuser', password='password', email='test@email.com')
+
+        self.article_site1 = ArticlePage(
+            title="Article 1 in site 1",
+            slug="article-1-in-site-1",
+            live=False,
+        )
+        self.yourmind.add_child(instance=self.article_site1)
+        self.article_site1.save_revision(
+            user=superuser, submitted_for_moderation=True)
+
+        self.article_site2 = ArticlePage(
+            title="Article 1 in site 2",
+            slug="article-1-in-site-2",
+            live=False,
+        )
+        self.yourmind2.add_child(instance=self.article_site2)
+        self.article_site2.save_revision(
+            user=superuser, submitted_for_moderation=True)
+
+        self.article2_site1 = ArticlePage(
+            title="Article 2 in site 1",
+            slug="article-2-in-site-1",
+            live=False,
+        )
+        self.yourmind.add_child(instance=self.article2_site1)
+        self.article2_site1.save_revision(
+            user=superuser, submitted_for_moderation=True)
+
+        self.article2_site2 = ArticlePage(
+            title="Article 2 in site 2",
+            slug="article-2-in-site-2",
+            live=False,
+        )
+        self.yourmind2.add_child(instance=self.article2_site2)
+        self.article2_site2.save_revision(
+            user=superuser, submitted_for_moderation=True)
+
+    def test_approve_moderation(self):
+        self.client.login(username='testuser', password='password')
+
+        self.assertFalse(self.article_site1.live)
+        self.assertTrue(
+            self.article_site1.revisions.first().submitted_for_moderation)
+        revision = self.article_site1.get_latest_revision()
+
+        response = self.client.post(reverse(
+            'wagtailadmin_pages:approve_moderation',
+            args=(revision.id,)))
+
+        self.assertRedirects(response, reverse('wagtailadmin_home'))
+        self.article_site1.refresh_from_db()
+        self.assertTrue(self.article_site1.live)
+        self.assertTrue(ArticlePage.objects.descendant_of(self.main).get(
+            slug=self.article_site1.slug))
+
+        self.assertFalse(self.article_site2.live)
+        self.assertTrue(
+            self.article_site2.revisions.first().submitted_for_moderation)
+        revision = self.article_site2.get_latest_revision()
+
+        response = self.client.post(reverse(
+            'wagtailadmin_pages:approve_moderation',
+            args=(revision.id,)))
+
+        self.assertRedirects(response, reverse('wagtailadmin_home'))
+        self.article_site2.refresh_from_db()
+        self.assertTrue(self.article_site2.live)
+        self.assertTrue(ArticlePage.objects.descendant_of(self.main2).get(
+            slug=self.article_site2.slug))
+
+    def test_reject_moderation(self):
+        self.client.login(username='testuser', password='password')
+        self.assertFalse(self.article2_site1.live)
+        self.assertTrue(
+            self.article2_site1.revisions.first().submitted_for_moderation)
+        revision = self.article2_site1.get_latest_revision()
+
+        response = self.client.post(reverse(
+            'wagtailadmin_pages:reject_moderation',
+            args=(revision.id, )))
+
+        self.assertRedirects(response, reverse('wagtailadmin_home'))
+        self.article2_site1.refresh_from_db()
+        self.assertFalse(self.article2_site1.live)
+        self.assertFalse(
+            self.article2_site1.revisions.first().submitted_for_moderation)
+        self.assertTrue(ArticlePage.objects.descendant_of(self.main).get(
+            slug=self.article2_site1.slug))
+
+        self.assertFalse(self.article2_site2.live)
+        self.assertTrue(
+            self.article2_site2.revisions.first().submitted_for_moderation)
+        revision = self.article2_site2.get_latest_revision()
+
+        response = self.client.post(reverse(
+            'wagtailadmin_pages:reject_moderation',
+            args=(revision.id, )))
+
+        self.assertRedirects(response, reverse('wagtailadmin_home'))
+        self.article2_site2.refresh_from_db()
+        self.assertFalse(self.article2_site2.live)
+        self.assertFalse(
+            self.article2_site2.revisions.first().submitted_for_moderation)
+        self.assertTrue(ArticlePage.objects.descendant_of(self.main2).get(
+            slug=self.article2_site2.slug))